requirements.txt
Package            Version
------------------ ---------
beautifulsoup4     4.10.0
certifi            2021.10.8
charset-normalizer 2.0.11
idna               3.3
joblib             1.1.0
numpy              1.22.2
pandas             1.4.1
pip                21.2.2
psycopg2           2.9.3
PyMySQL            1.0.2
python-dateutil    2.8.2
pytz               2021.3
requests           2.27.1
scikit-surprise    1.1.1
scipy              1.8.0
setuptools         58.0.4
six                1.16.0
soupsieve          2.3.1
urllib3            1.26.8
wheel              0.37.1
wincertstore       0.2
